Create header files to use in other cpp files
The reason that the all the implementations are also in (.h) files and not
in (.cpp) files is because most of the classes have template methods.
Which can't be compiled and linked like normal classes. Now, both of the
Graph classes do not have template methods, so they can be split into
one (.h) and one (.cpp) files. But I'm just lazy, and I do have plans to
add template methods in Graph classes.